// PetCloud/src/lib/db/schema/schema.ts
export const auth = pgSchema("auth");
// ...
export const usersInAuth = auth.table(
  "users",
  {
    id: uuid().defaultRandom().primaryKey().notNull(),
    aud: text().default("authenticated"),
    role: text().default("authenticated"),
    email: text(),
    encryptedPassword: text("encrypted_password"),
    emailConfirmedAt: timestamp("email_confirmed_at", {
      withTimezone: true,
      mode: "string",
    }),
    confirmedAt: timestamp("confirmed_at", {
      withTimezone: true,
      mode: "string",
    }),
    phone: text(),
    phoneConfirmedAt: timestamp("phone_confirmed_at", {
      withTimezone: true,
      mode: "string",
    }),
    rawAppMetaData: jsonb("raw_app_meta_data").default({}).notNull(),
    rawUserMetaData: jsonb("raw_user_meta_data").default({}).notNull(),
    bannedUntil: timestamp("banned_until", {
      withTimezone: true,
      mode: "string",
    }),
    lastSignInAt: timestamp("last_sign_in_at", {
      withTimezone: true,
      mode: "string",
    }),
    isAnonymous: boolean("is_anonymous").default(false).notNull(),
    isSsoUser: boolean("is_sso_user").default(false).notNull(),
    deletedAt: timestamp("deleted_at", { withTimezone: true, mode: "string" }),
    createdAt: timestamp("created_at", { withTimezone: true, mode: "string" })
      .defaultNow()
      .notNull(),
    updatedAt: timestamp("updated_at", { withTimezone: true, mode: "string" })
      .defaultNow()
      .notNull(),
  },
  (table) => [
    uniqueIndex("users_email_key")
      .using("btree", sql`lower(email)`)
      .where(sql`(email IS NOT NULL)`),
  ],
);